// src/app/chats/events/chats.events.enum.ts
export enum Event {
  RoomCreated = 'room-created',
  RoomUpdated = 'room-updated',
  RoomDeleted = 'room-deleted',
  MessageCreated = 'message-created',
  MessageDelivered = 'message-delivered',
  MessageRead = 'message-read',
  MessageEdited = 'message-edited',
  MessageDeleted = 'message-deleted',
  UserAdded = 'user-added',
  UserRemoved = 'user-removed',
}